+++ tests/test/hello_world.rs
@@ -69,7 +69,7 @@ fn hello_world()
 
     // launch websocket server
     tracing::info!("ws hello world test: launching server...");
-    let websocket_server = server_demo_factory().new_server(
+    let mut websocket_server = server_demo_factory().new_server(
             server_runtime,
             "127.0.0.1:0",
             bevy_simplenet::AcceptorConfig::Default,