+++ tests/test/rate_limit.rs
@@ -58,7 +58,7 @@ fn rate_limit_test(max_count_per_period: u32)
     let client_runtime = enfync::builtin::Handle::default();
 
     // launch websocket server
-    let websocket_server = server_demo_factory().new_server(
+    let mut websocket_server = server_demo_factory().new_server(
             server_runtime,
             "127.0.0.1:0",
             bevy_simplenet::AcceptorConfig::Default,