+++ tests/test/request_response.rs
@@ -63,7 +63,7 @@ fn request_response()
     let client_runtime = enfync::builtin::Handle::default();
 
     // launch websocket server
-    let websocket_server = server_demo_factory().new_server(
+    let mut websocket_server = server_demo_factory().new_server(
             server_runtime,
             "127.0.0.1:0",
             bevy_simplenet::AcceptorConfig::Default,
@@ -152,7 +152,7 @@ fn request_ack()
     let client_runtime = enfync::builtin::Handle::default();
 
     // launch websocket server
-    let websocket_server = server_demo_factory().new_server(
+    let mut websocket_server = server_demo_factory().new_server(
             server_runtime,
             "127.0.0.1:0",
             bevy_simplenet::AcceptorConfig::Default,
@@ -239,7 +239,7 @@ fn request_rejected()
     let client_runtime = enfync::builtin::Handle::default();
 
     // launch websocket server
-    let websocket_server = server_demo_factory().new_server(
+    let mut websocket_server = server_demo_factory().new_server(
             server_runtime,
             "127.0.0.1:0",
             bevy_simplenet::AcceptorConfig::Default,
@@ -326,7 +326,7 @@ fn request_dropped()
     let client_runtime = enfync::builtin::Handle::default();
 
     // launch websocket server
-    let websocket_server = server_demo_factory().new_server(
+    let mut websocket_server = server_demo_factory().new_server(
             server_runtime,
             "127.0.0.1:0",
             bevy_simplenet::AcceptorConfig::Default,